Add signed_request authenticator
The signed_request authenticator requires and Authorization header be
present of the form:
Authorization: SIGNED_REQUEST <SIGNATURE>.<PAYLOAD>
where signature contains a base64-encoded SHA256 HMAC signature of the
payload.  The payload is expected to be a base64-encoded JSON object
representing the user making the request, but the payload is not
currently used.